feat: Add more data to city paginated list
A new class PaginatedList that includes:

* rows: the previous list
* totalRecords: the number of total records
* pageSize: the page size
* page: the page
* query: an optional query str

Is returned, this model should be used for all paginated results.

We also re-add the city cache, this cache is split between the returned
page and the returned count, so if we are paginating the list, only the
first time the total size is calculated.

Signed-off-by: Arturo Volpe <arturovolpe@gmail.com>